Support voidwaker spec
closes #35

much more complicated than expected because of the magic typing, plus it's a re-used spec-animation (also dwh/swh spec) which brought in new types of exceptions. Kinda spaghetti at this point since these types of exceptions weren't foreseen, but it works. Tested in LMS, and with the help of random LMS lobby players 'ELL Cucuy' & 'DannyMaulz' to help test the voidwaker spec, since it's not easy to obtain in LMS.