Hits on robes: fix how calculateRobeHits is called
used to be in Overlay's .Render(), which is really inefficient and unnecessary, just needs to be called once per attack.

This still isn't optimal as it uses a loop when it could be basic ongoing counters, like most other statistics, but should be negligible in practice since the fight logs don't get that big due to how long fights last.